% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hue_ip.R
\name{get_hue_ip}
\alias{get_hue_ip}
\title{Get the internal IP address of your Hue bridge}
\usage{
get_hue_ip()
}
\description{
This function uses the bridge UPNP service to get the internal IP address
of your hue hub. You need to be on the same network, else you'll get an error.
This is not the only method to get the bridge's ip address, but is probably the easiest.
This function has no parameters/inputs.
}
\section{Bugs}{

Code repo: \url{https://github.com/epijim/notifyme}
}
\examples{
\dontrun{get_hue_ip()}
}
\keyword{Hue}
\keyword{R}
\keyword{notify}

