% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_light_info.R
\name{get_light_info}
\alias{get_light_info}
\title{Get info on lights connected to your Hue hub}
\usage{
get_light_info(bridge_ip = NULL, username = NULL,
  file = "~/r_keychain.rds")
}
\arguments{
\item{bridge_ip}{Internal IP address of your hue bridge}

\item{username}{Username for connecting to hue bridge}

\item{file}{If the file listed here exists, it will try and use the private keys created by the function save_private_keys(). Just ignore if you want to manually give your bridge ip and username.}
}
\description{
This will return a dataframe containing information about the lights
connected to your hue hub. It includes info on current, colour, is it on,
and is it powered/reeachable.
}
\section{Bugs}{

Code repo: \url{https://github.com/epijim/notifyme}
}
\examples{
\dontrun{get_light_info(bridge_ip,username)}
}
\keyword{Hue}
\keyword{R}
\keyword{notify}

