% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_private_keys.R
\name{get_private_keys}
\alias{get_private_keys}
\title{Load a key from the keychain made by save_private_keys()}
\usage{
get_private_keys(api_var = "pushover_userkey", file = "~/r_keychain.rds")
}
\arguments{
\item{api_var}{The name of the api key, this is user defined.}

\item{file}{The name and location of the file where you want to store it. Default is same as the save function.}
}
\description{
This function will look for a 'keychain' file, and if found load the key you asked for. Designed to
be used with save_private_keys()
}
\section{Bugs}{

Code repo: \url{https://github.com/epijim/notifyme}
}
\examples{
\dontrun{get_private_keys("keyImInterestedIn")}
}
\keyword{Hue}
\keyword{R}
\keyword{notify}
\keyword{storekeys}

